Extract parseHtml
